Moved scheduler to lpm/back-end
